rename: common 패키지 생성 및 response 이동
